pyproject.toml
[tool]
deps = ["torch", "transformers", "peft", "datasets", "trl", "scikit-learn"]

[project]
name = "your-novel-helper"
version = "0.1.0"
description = "A novel style fine-tuning project based on Qwen3-7B"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "datasets>=2.18.0",
    "trl>=0.8.0",
    "scikit-learn>=1.4.0",
    "accelerate>=0.28.0",
    "pyyaml>=6.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "gradio>=4.0.0",
    "jieba>=0.42.0",
    "tqdm>=4.66.0",
]
